split modules for cli and api
